Add :data_source to `beacon_site/2`
