+++ test/support/router.ex
@@ -13,6 +13,6 @@ defmodule Beacon.BeaconTest.Router do
   scope "/" do
     pipe_through :browser
     beacon_admin "/page_management"
-    beacon_site "/", name: "my_site"
+    beacon_site "/", name: "my_site", data_source: Beacon.BeaconTest.BeaconDataSource
   end
 end
